refactor: rename a directory
