+++ packages/safe-ds-lang/src/language/index.ts
@@ -4,7 +4,7 @@ import {
     RPC_RUNNER_START,
     RPC_RUNNER_STARTED,
     RPC_RUNNER_UPDATE,
-} from './runner/safe-ds-runner.js';
+} from './runtime/safe-ds-runner.js';
 
 // Services
 export type { SafeDsServices } from './safe-ds-module.js';
@@ -24,7 +24,7 @@ export * from './helpers/nodeProperties.js';
 export { locationToString, positionToString, rangeToString } from '../helpers/locations.js';
 
 // Messages
-export * as messages from './runner/messages.js';
+export * as messages from './runtime/messages.js';
 
 // Remote procedure calls
 export const rpc = {
